Solving icon click with pointerevents allows label to be extracted
If the label is integral to solving the icon-click because it wraps the whole component, then the label can't be moved out, or replaced by something outside the component without a high likelihood of re-introducing this bug.

Solving this with pointer-events: none, to allow a click through to the input that sits behind it seems more future proof.