Defining the custom variable inside the JSX
Creating the custom variable in the JSX, allows for the styles object data to be used for other purposes as well. In this case the iconSize and lineThickness can now easily be used for the icon properties.

calculating the actual rem (and px) values in the JSX is not strictly necessary, it can be done with calc(var(--foo) /16 * 1rem), but it reads cleaner like this.

Additional changes:
- width can be set by the input
- only padding-left on the input is really necessary (which is the same as the input height) unless even more control on text position is needed.